[#18] refactor > Input 컴포넌트 분리 (UserPageInput)
