+++ client/src/components/Join/index.tsx
@@ -3,6 +3,7 @@ import { Form, Field } from 'react-final-form';
 import { createUserValidation } from '../../utils/validation';
 import * as S from './style';
 import { UserJoin } from '../../types/data';
+import UserPageInput from '../UserPageInput';
 
 interface Props {
   onSubmit: (values: UserJoin) => void;
@@ -20,55 +21,35 @@ export default function Join({ onSubmit }: Props) {
             <>
               <Field name="email">
                 {({ input, meta }) => (
-                  <S.InputWrapper>
-                    <S.Input {...input} type="email" placeholder="이메일" />
-                    <S.InputError visible={meta.touched && meta.error ? true : false}>
-                      {meta.error}
-                    </S.InputError>
-                  </S.InputWrapper>
+                  <UserPageInput input={input} meta={meta} type="email" placeholder="이메일" />
                 )}
               </Field>
               <Field name="name">
                 {({ input, meta }) => (
-                  <S.InputWrapper>
-                    <S.Input {...input} type="text" placeholder="이름" />
-                    <S.InputError visible={meta.touched && meta.error ? true : false}>
-                      {meta.error}
-                    </S.InputError>
-                  </S.InputWrapper>
+                  <UserPageInput input={input} meta={meta} type="text" placeholder="이름" />
                 )}
               </Field>
               <Field name="nickname">
                 {({ input, meta }) => (
-                  <S.InputWrapper>
-                    <S.Input {...input} type="nickname" placeholder="닉네임" />
-                    <S.InputError visible={meta.touched && meta.error ? true : false}>
-                      {meta.error}
-                    </S.InputError>
-                  </S.InputWrapper>
+                  <UserPageInput input={input} meta={meta} type="text" placeholder="닉네임" />
                 )}
               </Field>
               <Field name="password">
                 {({ input, meta }) => (
-                  <S.InputWrapper>
-                    <S.Input {...input} type="password" placeholder="비밀번호" />
-                    <S.InputError visible={meta.touched && meta.error ? true : false}>
-                      {meta.error}
-                    </S.InputError>
-                  </S.InputWrapper>
+                  <UserPageInput input={input} meta={meta} type="password" placeholder="비밀번호" />
                 )}
               </Field>
               <Field name="passwordConfirm">
                 {({ input, meta }) => (
-                  <S.InputWrapper>
-                    <S.Input {...input} type="password" placeholder="비밀번호 확인" />
-                    <S.InputError visible={meta.touched && meta.error ? true : false}>
-                      {meta.error}
-                    </S.InputError>
-                  </S.InputWrapper>
+                  <UserPageInput
+                    input={input}
+                    meta={meta}
+                    type="password"
+                    placeholder="비밀번호 확인"
+                  />
                 )}
               </Field>
-              <S.PushButton onClick={() => handleSubmit()}>arrow_right</S.PushButton>
+              <S.PushButton onClick={handleSubmit}>arrow_right</S.PushButton>
             </>
           )}
         />
