+++ client/src/components/Login/index.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import * as S from './style';
+
+export default function Login() {
+  return <div>123</div>;
+}
