+++ client/src/components/UserPageHeader/index.tsx
@@ -7,12 +7,12 @@ export default function UserPageHeader({ subPath }: { subPath: SubPath }) {
   return (
     <S.Wrapper>
       <S.NavigationWrapper>
-        <S.Navigation to="/user/login" isActive={subPath === 'login'}>
-          Sign In
-        </S.Navigation>
-        <S.Navigation to="/user/join" isActive={subPath === 'join'}>
-          Sign Up
-        </S.Navigation>
+        <S.NavigationBox isActive={subPath === 'login'}>
+          <S.Navigation to="/user/login">Sign In</S.Navigation>
+        </S.NavigationBox>
+        <S.NavigationBox isActive={subPath === 'join'}>
+          <S.Navigation to="/user/join">Sign Up</S.Navigation>
+        </S.NavigationBox>
       </S.NavigationWrapper>
     </S.Wrapper>
   );
